combine eval commands
