+++ extension.js
@@ -1,7 +1,6 @@
 const vscode = require('vscode');
-const { stringifyError } = require('./util');
 const Lang = require('supercolliderjs').lang.default;
-const { flashHighlight } = require('./util');
+const { evalHighlighted, evalRegion } = require('./evaluate');
 
 let lang = null;
 
@@ -133,8 +132,8 @@ async function activate(context) {
 
   context.subscriptions.push(hushAll);
 
-  const evaluateHighlighted = vscode.commands.registerCommand(
-    'supercollider.evaluateHighlighted',
+  const evaluate = vscode.commands.registerCommand(
+    'supercollider.evaluate',
     async () => {
       if (!lang) {
         console.error('sclang not started, cannot boot scsynth.');
@@ -143,118 +142,16 @@ async function activate(context) {
 
       const editor = vscode.window.activeTextEditor;
       const selection = editor.selection;
+
       if (selection && !selection.isEmpty) {
-        const selectionRange = new vscode.Range(
-          selection.start.line,
-          selection.start.character,
-          selection.end.line,
-          selection.end.character
-        );
-        const highlighted = editor.document.getText(selectionRange);
-
-        try {
-          const result = await lang.interpret(highlighted, null, true, false);
-          postWindow.appendLine(result.trim());
-          flashHighlight(vscode.window.activeTextEditor, selectionRange);
-        } catch (err) {
-          postWindow.appendLine(err);
-          console.error(err);
-        }
+        evalHighlighted(lang, vscode, postWindow, editor, selection);
+      } else {
+        evalRegion(lang, vscode, postWindow, editor, hyperScopes);
       }
     }
   );
 
-  context.subscriptions.push(evaluateHighlighted);
-
-  const evalRegion = vscode.commands.registerCommand(
-    'supercollider.evalRegion',
-    async () => {
-      if (!lang) {
-        postWindow.appendLine('sclang not started, cannot evaluate region.');
-        return;
-      }
-
-      const editor = vscode.window.activeTextEditor;
-      const ranges = [];
-      let brackets = 0;
-
-      // Get the total line count of the open script
-      const lineCount = vscode.window.activeTextEditor.document.lineCount;
-
-      // For every line, get the text of that line as a string.
-      for (let i = 0; i < lineCount; i++) {
-        const { text } = vscode.window.activeTextEditor.document.lineAt(i);
-
-        // for every character on the line, check to see if it's a paren
-        for (let j = 0; j < text.length; j++) {
-          const char = text.charAt(j);
-
-          // not totally sure about this --
-          // it has been hastily ported it from hadron editor so I still gotta learn how it works.
-          const { scopes } = hyperScopes.getScopeAt(
-            editor.document,
-            new vscode.Position(i, j)
-          );
-          const scopesLength = scopes.length - 1;
-          const lastScope = scopes[scopesLength];
-          if (
-            lastScope === 'comment.single.supercollider' ||
-            lastScope === 'comment.multiline.supercollider' ||
-            lastScope === 'string.quoted.double.supercollider' ||
-            lastScope === 'entity.name.symbol.supercollider' ||
-            lastScope === 'constant.character.escape.supercollider'
-          ) {
-            continue;
-          }
-
-          // gather the bracket ranges
-          if (char === '(' && brackets++ === 0) {
-            ranges.push([i]);
-          } else if (char === ')' && --brackets === 0) {
-            ranges[ranges.length - 1].push(i);
-          }
-        }
-      }
-      // Get where the current cursor is
-      const position = vscode.window.activeTextEditor.selection.active;
-
-      // not totally sure about this --
-      // it has been hastily ported it from hadron editor so I still gotta learn how it works.
-      const range = ranges.find((range) => {
-        return range[0] <= position.c && position.c <= range[1];
-      });
-
-      const { text } = vscode.window.activeTextEditor.document.lineAt(range[1]);
-      const lastLineLength = text.length;
-
-      const selectionRange = new vscode.Range(
-        range[0],
-        0, // will need to calculate ... because what if the paren is not the first char on the line? does that matter?
-        range[1],
-        lastLineLength
-      );
-      const highlighted = editor.document.getText(selectionRange);
-
-      try {
-        const result = await lang.interpret(
-          highlighted,
-          null,
-          true,
-          true,
-          true
-        );
-        console.log(result);
-        postWindow.appendLine(result.trim());
-        flashHighlight(vscode.window.activeTextEditor, selectionRange);
-      } catch (err) {
-        const errString = stringifyError(err);
-        postWindow.appendLine(errString);
-        console.error(errString);
-      }
-    }
-  );
-
-  context.subscriptions.push(evalRegion);
+  context.subscriptions.push(evaluate);
 }
 
 exports.activate = activate;
